Problems #18 and #67
These are solved using top down (greedy algorithm), as the problem
states it should be done, however, the actual answer to the problem is
achieved using bottom up (dynamic programming). Therefore, I have left
the mystery of the true solution to someone else, but it is in essence
the same problem as given here but starting at the bottom, finding the
max, and moving up the triangle.